Update AppContext.js
Here, we are importing AppContext from your Context.Import the useContext hook, and pass your AppContext to it - this is how a component connects to the context in order to get values from global state. The Bootstrap Alert classes are used to give a nice gray background by adding some text and hard coding a value.